PCB: Reposition all sensor headers near edge
Since headers will have cables running off into the sensors themselves
(aside from temp), their respective headers should be at the edges of
the board. The raindrop sensor, however, is a breakout board itself with
its own header, so it has been placed so that its external header is
facing the edge of the board as well.