App: Remove unused area unit choices
Area can be only entered in as either square meters or square feet
because anything smaller doesn't make sense for the irrigation
algorithm. The area is always written to the characteristic as square
meters, however, and so it is converted first if entered as square feet.

The smallest allowable area is 1 square meter and anything smaller will
be rounded up to 1.