Add ability to download stored data onto phone
The phone is able to alert the device that it wishes to download any
stored data with the History Signal characteristic. When the device
receives that signal it begins to write data to the History
characteristic (if any such data exists). The phone is alerted when that
characterisitc has been updated via the Indicate property and it writes
the data to a timestamped CSV file. The device gets alerted that the
phone has read the data and it writes another piece to the
characteristic. This happens until all data has been stored by the
phone.